+++ ChoGGi's Library/Changelog.md
@@ -6,6 +6,7 @@
 - Made X close icon a little smaller, so it fits the height of the Examine toolbar.
 - Examine: "Go to text" is now "Search".
 - More params for OpenInMultiLineTextDlg.
+- Sorts list of mods in Options>Mod Options.
 
 ## v9.7 (10 Apr 2021)
 - Added IsArray.